+++ app/Http/Controllers/ProyekController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Milestone;
 use App\Models\Proyek;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class ProyekController extends Controller
 {
@@ -29,9 +30,11 @@ class ProyekController extends Controller
         if(request('query')){
             $proyek = Proyek::where('nama_proyek', 'LIKE', '%' . request('query') . '%')
                         ->orWhere('lokasi', 'LIKE', '%' . request('query') . '%')
+                        ->orderBy('tgl_mulai_proyek', 'DESC')
                         ->paginate($limit);
         }else{
-            $proyek = Proyek::paginate($limit);
+            $proyek = Proyek::orderBy('tgl_mulai_proyek', 'DESC')
+                        ->paginate($limit);
         }
         return response()->json($proyek);
     }
@@ -74,18 +77,26 @@ class ProyekController extends Controller
         $proyek->harga = $request['harga'];
         $proyek->tgl_mulai_proyek = $request['tgl_mulai_proyek'];
         $proyek->tgl_selesai_proyek = $request['tgl_selesai_proyek'];
-        $proyek->save();
-        $proyek->id_proyek;
-
-        $milestone = new Milestone();
-        $milestone->nama_milestone = $proyek->nama_proyek;
-        $milestone->id_proyek = $proyek->id_proyek;
-        $milestone->save();
-
-        $data = [
-            $proyek->save(),
-            $milestone->save()
-        ];
+        $tgl_finish = date('Y-m-d', strtotime($proyek->tgl_selesai_proyek));
+        $tgl_start = date('Y-m-d', strtotime($proyek->tgl_mulai_proyek));
+        if(strtotime($tgl_start) > strtotime($tgl_finish)){
+            $data = "Maaf tanggal mulai harus lebih kecil dari tanggal selesai";
+            echo "$data";
+        }
+        else{
+            $proyek->save();
+            $proyek->id_proyek;
+    
+            $milestone = new Milestone();
+            $milestone->nama_milestone = $proyek->nama_proyek;
+            $milestone->id_proyek = $proyek->id_proyek;
+            $milestone->save();
+    
+            $data = [
+                $proyek->save(),
+                $milestone->save()
+            ];
+        }
 
         // $add = Proyek::create($request->all());
         return response()->json($data);
@@ -143,9 +154,28 @@ class ProyekController extends Controller
         $proyek->harga = $data['harga'];
         $proyek->tgl_mulai_proyek = $data['tgl_mulai_proyek'];
         $proyek->tgl_selesai_proyek = $data['tgl_selesai_proyek'];
-        $proyek->save();
+        $tgl_finish = date('Y-m-d', strtotime($proyek->tgl_selesai_proyek));
+        $tgl_start = date('Y-m-d', strtotime($proyek->tgl_mulai_proyek));
+        if(strtotime($tgl_start) > strtotime($tgl_finish)){
+            $data = "Maaf tanggal mulai harus lebih kecil dari tanggal selesai";
+            echo "$data";
+        }
+        else{
+            $proyek->save();
+            $proyek->id_proyek;
+    
+            $milestone = Milestone::find($id);
+            $milestone->nama_milestone = $proyek->nama_proyek;
+            $milestone->id_proyek = $proyek->id_proyek;
+            $milestone->save();
+    
+            $data = [
+                $proyek->save(),
+                $milestone->save()
+            ];
+        }
 
-        return response()->json($proyek);
+        return response()->json($data);
     }
 
     /**
@@ -156,7 +186,25 @@ class ProyekController extends Controller
      */
     public function destroy($id)
     {
-        $hapus = Proyek::destroy($id);
+        $hasil = DB::select('select * from milestone where id_proyek = ?', [$id]);
+        if(!empty($milestone)){
+            foreach($milestone as $m){
+                $mlst = [
+                    'id_proyek' => $m->id_proyek,
+                    'nama_milestone' => $m->nama_milestone
+                ];
+                $tgs = DB::select('select * from tugas where id_milestone = ?', [$m->id_milestone]);
+                if(!empty($tgs)){
+                    $hapus = "Proyek Tidak Dapat Dihapus, Dikarenakan Terdapat Tugas Didalamnya!";
+                    echo "$hapus";
+                }else{
+                    $hapus = Proyek::destroy($id);
+                }
+            }
+            return response()->json($hapus);
+        }else{
+            $hapus = Proyek::destroy($id);
+        }
         return response()->json($hapus);
     }
 }
